Added new algorithm and fixed the old one
Added Genetic Algorithm with Simulated Annealing.
In Genetic Algorithm changed thrown out graph as a field of class
Individual and changed the way of loading into fields of that class, and
fixed logic of some other classes so the same edge representet on two
different ways doesn't  cause the problem.
Each algorithm is inserted in its own directory.